[Win]  Break out editor automation peer for events
