+++ Xamarin.PropertyEditing.Windows/EventEditor.cs
@@ -1,9 +1,5 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
+using System.Windows.Automation.Peers;
 using System.Windows.Controls;
 using System.Windows.Input;
 
@@ -37,6 +33,11 @@ namespace Xamarin.PropertyEditing.Windows
 	    {
 		    get { return (DataTemplate) GetValue (ItemTemplateProperty); }
 		    set { SetValue (ItemTemplateProperty, value); }
+	    }
+
+	    protected override AutomationPeer OnCreateAutomationPeer ()
+	    {
+		    return new EditorAutomationPeer (this);
 	    }
     }
 }
